% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.collect.R
\name{degree.collect}
\alias{degree.collect}
\title{Degrees of nodes}
\usage{
degree.collect(net)
}
\arguments{
\item{net}{The input network.}
}
\value{
A vector.
}
\description{
Collect the degrees for all nodes in a network.
}
\details{
Obtain the degrees for all nodes.
}
\examples{
\dontrun{
x.deg <- degree.collect(x)
summary(x.deg)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
