% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.hist.R
\name{degree.hist}
\alias{degree.hist}
\title{Histogram of the degree distribution of a network}
\usage{
degree.hist(g, breaks = 100)
}
\arguments{
\item{g}{The input network.}

\item{breaks}{A single number giving the number of cells for the histogram.}
}
\value{
A .gif plot.
}
\description{
Plot the histogram of all degrees of a network.
}
\details{
Plot the histogram of all degrees of a network.
}
\examples{
\dontrun{
x <-  net.erdos.renyi.gnp(1000, 0.05)
degree.hist(x)}
}
\author{
Xu Dong
}
