% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.deg.R
\name{dist.deg}
\alias{dist.deg}
\title{Degree Distribution}
\usage{
dist.deg(net)
}
\arguments{
\item{net}{The input network.}
}
\value{
A vector.
}
\description{
Obtain the degrees for all nodes.
}
\details{
Obtain the degrees for all nodes.
}
\examples{
\dontrun{
x.dist.deg <- dist.deg(x)
summary(x.dist.deg)}
}
\author{
Xu Dong, Nazrul Shaikh.
}

