% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.degdist.cum.R
\name{draw.degdist.cum}
\alias{draw.degdist.cum}
\title{Plot of the cumulative degree distribution of a network}
\usage{
draw.degdist.cum(net)
}
\arguments{
\item{net}{The input network.}
}
\value{
A .gif plot.
}
\description{
Plot the cumulative degree distribution of a network, with a logarithm scale.
}
\details{
Plot the cumulative degree distribution of a network, with a logarithm scale.
}
\examples{
\dontrun{
x <-  net.barabasi.albert(1000, 20)
draw.degdist.cum(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
