% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.degdist.log.R
\name{draw.degdist.log}
\alias{draw.degdist.log}
\title{Plot of the degree distribution of a network, with a logarithm scale}
\usage{
draw.degdist.log(net)
}
\arguments{
\item{net}{The input network.}
}
\value{
A .gif plot.
}
\description{
Plot the degree distribution of a network, with a logarithm scale.
}
\details{
Plot the degree distribution of a network, with a logarithm scale.
}
\examples{
\dontrun{
x <-  net.erdos.renyi.gnp(1000, 0.01)
draw.degdist.log(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}

