% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.net.R
\name{draw.net}
\alias{draw.net}
\title{Plot of a small network}
\usage{
draw.net(net)
}
\arguments{
\item{net}{The input network.}
}
\value{
A .gif plot.
}
\description{
Plot a small network.
}
\details{
Plot a small network.
}
\examples{
\dontrun{
x <-  net.ring.lattice(12,4)
draw.net(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}

