% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.adjacency.R
\name{from.adjacency}
\alias{from.adjacency}
\title{Adjacency Matrix to \code{fastnet}}
\usage{
from.adjacency(adj.mat)
}
\arguments{
\item{adj.mat}{The input adjacency matrix}
}
\value{
A list containing the nodes of the network and their respective neighbors.
}
\description{
Transform an adjacency matrix to an ego-centric list form used in \code{fastnet}.
}
\examples{
adj.mat <- matrix(c(0,1,0,0,1,0,0,0,0,0,0,1,0,0,1,0), nrow = 4, ncol = 4)
g <- from.adjacency(adj.mat)

}
\author{
Xu Dong, Christian Llano.
}
