% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.csc.R
\name{from.csc}
\alias{from.csc}
\title{Sparse .csc Adjacency Matrix to \code{fastnet}}
\usage{
from.csc(sparse.mat.csc)
}
\arguments{
\item{sparse.mat.csc}{The input adjacency matrix stored in the .csc sparse matrix form}
}
\value{
A list containing the nodes of the network and their respective neighbors
}
\description{
Transform a sparse .csc adjacency matrix to an ego-centric list form used in \code{fastnet}.
}
\author{
Xu Dong
}
