% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.csr.R
\name{from.csr}
\alias{from.csr}
\title{Sparse .csr Adjacency Matrix to \code{fastnet}}
\usage{
from.csr(sparse.mat.csr)
}
\arguments{
\item{sparse.mat.csr}{The input adjacency matrix stored in the .csr sparse matrix form}
}
\value{
A list containing the nodes of the network and their respective neighbors
}
\description{
Transform a sparse .csr adjacency matrix to an ego-centric list form used in \code{fastnet}.
}
\author{
Xu Dong
}
