% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.statnet.R
\name{from.statnet}
\alias{from.statnet}
\title{\code{statnet}  to \code{fastnet}}
\usage{
from.statnet(net.statnet, ncores = detectCores())
}
\arguments{
\item{net.statnet}{The input \code{statnet} object.}

\item{ncores}{The number of cores to be used.}
}
\value{
A list containing the nodes of the network and their respective neighbors.
}
\description{
Import a \code{statnet} object.
}
\examples{
\dontrun{
library("ergm")
library("doParallel")
data("flo")
nflo <- network(flo, loops = TRUE)
fflo <- from.statnet(nflo)}
}
\author{
Xu Dong
}
