% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.neighbors.R
\name{get.neighbors}
\alias{get.neighbors}
\title{Neighbors of an agent in a network}
\usage{
get.neighbors(net, NodeID)
}
\arguments{
\item{net}{The input network.}

\item{NodeID}{The ID of the input node.}
}
\value{
A vector.
}
\description{
Presents all neighbors of a given node.
}
\details{
Neighbors of a node are nodes that directly connects to this node.
}
\examples{
\dontrun{
x <-  net.ring.lattice(12,4)
get.neighbors(x,2)}
}
\author{
Xu Dong
}
