% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo.dist.R
\name{histo.dist}
\alias{histo.dist}
\title{Histogram of all degrees of a network}
\usage{
histo.dist(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A .gif plot.
}
\description{
Plot the histogram of all degrees of a network.
}
\details{
Plot the histogram of all degrees of a network.
}
\examples{
\dontrun{
x <-  net.erdos.renyi.gnp(1000, 0.05)
histo.dist(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
