% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.degree.effective.R
\name{metric.degree.effective}
\alias{metric.degree.effective}
\title{Effective Degree}
\usage{
metric.degree.effective(g, effective_rate = 0.9)
}
\arguments{
\item{g}{The input network.}

\item{effective_rate}{The effective rate (0.9 is set by default).}
}
\value{
A real constant.
}
\description{
Calculate the effective degree of a network.
}
\details{
The effective degree
}
\examples{
\dontrun{
metric.degree.effective(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}

