% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.degree.max.efficient.R
\name{metric.degree.max.efficient}
\alias{metric.degree.max.efficient}
\title{Efficient Maximal Degree}
\usage{
metric.degree.max.efficient(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A real constant.
}
\description{
Calculate the efficient maximal degree of a graph.
}
\details{
The efficient maximal degree is the 90% quantile of the degrees of all nodes in graph g.
}
\examples{
\dontrun{x <-  net.erdos.renyi.gnp(1000, 0.01)
metric.degree.max.efficient(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
