% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.degree.max.R
\name{metric.degree.max}
\alias{metric.degree.max}
\title{Maximal Degree}
\usage{
metric.degree.max(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A real constant.
}
\description{
Calculate the maximal degree of a graph.
}
\details{
The maximal degree.
}
\examples{
\dontrun{
metric.degree.max(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}

