% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.degree.mean.R
\name{metric.degree.mean}
\alias{metric.degree.mean}
\title{Mean Degree}
\usage{
metric.degree.mean(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A real constant.
}
\description{
Calculate the mean degree of a graph.
}
\details{
The mean degree is the average value of the degrees of all nodes in graph g.
}
\examples{
\dontrun{
x <-  net.erdos.renyi.gnp(1000, 0.01)
metric.degree.mean(x)
}
}
\author{
Xu Dong, Nazrul Shaikh.
}
