% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.degree.median.R
\name{metric.degree.median}
\alias{metric.degree.median}
\title{Median Degree}
\usage{
metric.degree.median(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A real constant.
}
\description{
Calculate the median degree of a graph.
}
\details{
The median degree is the median value of the degrees of all nodes in graph g.
}
\examples{
\dontrun{
x <-  net.erdos.renyi.gnp(1000, 0.01)
metric.degree.median(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}

