% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.degree.min.R
\name{metric.degree.min}
\alias{metric.degree.min}
\title{Minimal Degree}
\usage{
metric.degree.min(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A real constant.
}
\description{
Calculate the minimal degree of a network.
}
\details{
The minimal degree.
}
\examples{
\dontrun{
metric.degree.min(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
