% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.degree.sd.R
\name{metric.degree.sd}
\alias{metric.degree.sd}
\title{Standard Deviation of Degree Distribution}
\usage{
metric.degree.sd(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A real constant.
}
\description{
Calculate the standard deviation of all degrees of a network.
}
\details{
The standard deviation of all degrees of a network.
}
\examples{
\dontrun{
metric.degree.sd(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
