% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.complete.R
\name{net.complete}
\alias{net.complete}
\title{Complete Network}
\usage{
net.complete(n, ncores = detectCores())
}
\arguments{
\item{n}{Number of nodes of the network.}

\item{ncores}{Number of cores, by default \code{detectCores()} from \code{parallel}.}
}
\value{
A list containing the nodes of the network and their respective neighbors.
}
\description{
Simulate a complete (or full) network.
}
\details{
The \emph{n} nodes in the network are fully connected.

Note that the input \emph{n} should not excess 10000, for the sake of memory overflow.
}
\examples{
\dontrun{
x <- net.complete(1000) #using ncores by default}
}
\author{
Xu Dong, Nazrul Shaikh.
}

