% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.deg.R
\name{preview.deg}
\alias{preview.deg}
\title{Preview of the degree distribution of a network}
\usage{
preview.deg(g)
}
\arguments{
\item{g}{The input network.}
}
\value{
A vector.
}
\description{
Present the first 10 degrees of a network.
}
\details{
Present the first 10 degrees of a network.
}
\examples{
\dontrun{
x <-  net.ring.lattice(12,4)
preview.deg(x)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
