% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.edgelist.R
\name{to.edgelist}
\alias{to.edgelist}
\title{\code{fastnet} to edgelist}
\usage{
to.edgelist(network, ncores)
}
\arguments{
\item{network}{A fastnet object.}

\item{ncores}{The number of cores to be used.}
}
\value{
A 2-column list with each row representing an edge, from source to destination
}
\description{
Coerce a fastnet object to edgelist.
}
\examples{
\dontrun{
g <- net.erdos.renyi.gnp(100, 0.1)
el <- to.edgelist(g)}
}
\author{
Xu Dong
}
