% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.igraph.R
\name{to.igraph}
\alias{to.igraph}
\title{\code{fastnet} to \code{igraph}}
\usage{
to.igraph(g)
}
\arguments{
\item{g}{A fastnet object}
}
\value{
An igraph object
}
\description{
Coerce a fastnet object to an igraph object
}
\author{
Xu Dong
}
