% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.tidygraph.R
\name{to.tidygraph}
\alias{to.tidygraph}
\title{\code{fastnet} to \code{tidygraph}}
\usage{
to.tidygraph(g)
}
\arguments{
\item{g}{A fastnet object.}
}
\value{
A tidygraph object
}
\description{
Coerce a fastnet object to a tidygraph object.
}
\author{
Xu Dong
}
