% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors}
\alias{neighbors}
\title{neighbors of an agent in a network}
\usage{
neighbors(net, node)
}
\arguments{
\item{net}{The input network.}

\item{node}{The input node.}
}
\value{
A vector.
}
\description{
Present all neighbors of a give agent in a network.
}
\details{
Neighbors of a node are nodes that directly connects to this node.
}
\examples{
\dontrun{
x <-  net.ring.lattice(12,4)
neighbors(x,2)}
}
\author{
Xu Dong, Nazrul Shaikh.
}

